refactor form -> common
