+++ frontend/src/components/Common/mixins.js
@@ -46,7 +46,7 @@ export const CodeLike = css`
 export const ButtonLike = css`
   font-family: 'JetBrains Mono', serif;
   font-weight: 700;
-  padding: calc(0.6em + 1px) 1.5em;
+  padding: 0.6em 1.5em;
   margin: 2em 0;
   outline: 0;
 `
